refactor: reorganize project structure and move utility packages
- Moved internal utility packages (formatter, response, time, validation) to the root pkg directory
- Updated import paths across multiple files to reflect new package locations
- Added new migration script for database management
- Simplified .gitignore to reflect new project structure
