+++ internal/repository/sql/item.go
@@ -4,8 +4,8 @@ import (
 	"database/sql"
 	"fmt"
 	"gocasts/ToDoApp/internal/model"
-	now "gocasts/ToDoApp/internal/pkg/time"
 	"gocasts/ToDoApp/internal/repository"
+	now "gocasts/ToDoApp/pkg/time"
 	"strconv"
 	"strings"
 	"time"
